starting re-linting and code cleanup
